feat(radio): migrate FrSkyBarData, CustomFunctionData, widget data, InputMixGroupBase to SourceRef (plan items 1-4)
1. FrSkyBarData.source / FrSkyLineData.sources → SourceRef
   Telemetry screen source fields now store SourceRef directly.

2. CustomFunctionData.all.val → LSValue
   Polymorphic val field now uses LSValue union (SourceRef/SwitchRef/int32_t).
   CFN_PARAM(p) returns .value (int32_t). CFN_PARAM_SRC(p) returns .source.

3. Widget option data → memcpy for SourceRef
   gauge.cpp and value.cpp read widget source options via memcpy.

4. InputMixGroupBase idx → SourceRef
   Constructor takes SourceRef directly. Callers construct
   SourceRef{SOURCE_TYPE_INPUT/CHANNEL, 0, n} instead of MIXSRC_* integers.

All 3 targets build.

+++ radio/src/edgetx.cpp
@@ -2126,7 +2126,7 @@ bool validateSFGV(CustomFunctionData* cfn)
   bool rv = false;
 
   if (CFN_FUNC(cfn) == FUNC_ADJUST_GVAR && CFN_GVAR_MODE(cfn) == FUNC_ADJUST_GVAR_CONSTANT) {
-    int16_t v = CFN_PARAM(cfn);
+    int32_t v = CFN_PARAM(cfn);
     int16_t vmin, vmax;
     getMixSrcRange(CFN_GVAR_INDEX(cfn) + MIXSRC_FIRST_GVAR, vmin, vmax);
     if (v < vmin) v = vmin;
